refactor: Move cURL converter to frontend (#11432)
