+++ packages/cli/package.json
@@ -112,7 +112,6 @@
     "convict": "6.2.4",
     "cookie-parser": "1.4.7",
     "csrf": "3.1.0",
-    "curlconverter": "3.21.0",
     "dotenv": "8.6.0",
     "express": "4.21.1",
     "express-async-errors": "3.1.1",
