--- packages/cli/src/curlconverter.d.ts
@@ -1,3 +1,0 @@
-declare module 'curlconverter' {
-	export function toJsonString(data: string): string;
-}
