+++ packages/editor-ui/package.json
@@ -56,6 +56,7 @@
     "change-case": "^5.4.4",
     "chart.js": "^4.4.0",
     "codemirror-lang-html-n8n": "^1.0.0",
+    "curlconverter": "^4.11.0",
     "comlink": "^4.4.1",
     "core-js": "^3.40.0",
     "dateformat": "^3.0.3",
@@ -89,6 +90,7 @@
     "vue-router": "catalog:frontend",
     "vue-virtual-scroller": "2.0.0-beta.8",
     "vue3-touch-events": "^4.1.3",
+    "web-tree-sitter": "0.24.3",
     "vuedraggable": "4.1.0",
     "xss": "catalog:"
   },
@@ -115,6 +117,7 @@
     "unplugin-icons": "^0.19.0",
     "unplugin-vue-components": "^0.27.2",
     "vite": "catalog:frontend",
+    "vite-plugin-static-copy": "2.2.0",
     "vite-svg-loader": "5.1.0",
     "vitest": "catalog:frontend",
     "vitest-mock-extended": "catalog:frontend",
